Initial commit for a Latin Study Aid using the CLTK.
The program accepts an input of raw text in Latin and returns a list of each word with a corresponding POS tag, as well as a list of all original words (with stopwords filtered) and their lemmata.

The next goal is to provide fuctionality to look up each lemma and provide definitions.